Progress on rewriting Sonos provider.
